+++ src/main/resources/lib/ssb/dataset/publishOld.ts
@@ -3,7 +3,8 @@ import { get as getContent, Content } from '/lib/xp/content'
 import { NodeQueryResultHit } from '/lib/xp/node'
 import { send } from '/lib/xp/event'
 import { executeFunction, sleep } from '/lib/xp/task'
-import { isSameOrBefore, isSameDay } from '/lib/ssb/utils/dateUtils'
+import { isSameOrBefore } from '/lib/ssb/utils/dateUtils'
+import { isSameDay } from '/lib/vendor/dateFns'
 
 import { Events, logUserDataQuery } from '/lib/ssb/repo/query'
 import { getDataSourceIdsFromStatistics, getStatisticsContent } from '/lib/ssb/dashboard/statistic'
