+++ src/main/resources/lib/ssb/utils/dateUtils.ts
@@ -1,36 +1,17 @@
-import { type Locale } from 'date-fns'
-import { default as nb } from 'date-fns/locale/nb'
-import { default as nn } from 'date-fns/locale/nn'
-import { default as enGB } from 'date-fns/locale/en-GB'
-import { default as parseISO } from 'date-fns/parseISO'
-import { default as format } from 'date-fns/format'
-import { default as isWithinInterval } from 'date-fns/isWithinInterval'
-import { default as addDays } from 'date-fns/addDays'
-import { default as subDays } from 'date-fns/subDays'
-import { default as getMonth } from 'date-fns/getMonth'
-import { default as getYear } from 'date-fns/getYear'
-import { default as getDate } from 'date-fns/getDate'
-import { default as isAfter } from 'date-fns/isAfter'
-import { default as isBefore } from 'date-fns/isBefore'
-import { default as isSameDay } from 'date-fns/isSameDay'
-import { default as isSameSecond } from 'date-fns/isSameSecond'
-import { default as formatDistanceToNowStrict } from 'date-fns/formatDistanceToNowStrict'
-import { formatDate as libTimeFormatDate } from '/lib/time'
-
-export {
+import {
+  Locale,
+  nb,
+  nn,
+  enGB,
   parseISO,
-  addDays,
-  subDays,
-  getMonth,
   format,
-  getYear,
-  getDate,
-  isAfter,
   isBefore,
   isSameDay,
   isSameSecond,
   isWithinInterval,
-}
+  formatDistanceToNowStrict,
+} from '/lib/vendor/dateFns'
+import { formatDate as libTimeFormatDate } from '/lib/time'
 
 export function sameDay(d1: Date, d2: Date): boolean {
   return d1.getDate() === d2.getDate() && d1.getMonth() === d2.getMonth() && d1.getFullYear() === d2.getFullYear()
