+++ src/main/resources/services/news/news.ts
@@ -1,7 +1,8 @@
 import { query, Content } from '/lib/xp/content'
 import { pageUrl } from '/lib/xp/portal'
 import { StatisticInListing, VariantInListing } from '/lib/ssb/dashboard/statreg/types'
-import { subDays, isSameDay, format, parseISO, getTimeZoneIso } from '/lib/ssb/utils/dateUtils'
+import { getTimeZoneIso } from '/lib/ssb/utils/dateUtils'
+import { subDays, isSameDay, format, parseISO } from '/lib/vendor/dateFns'
 import { fetchStatisticsWithReleaseToday } from '/lib/ssb/statreg/statistics'
 import { isEnabled } from '/lib/featureToggle'
 import { xmlEscape } from '/lib/text-encoding'
